% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/are_exosplice_vars.R
\name{are_exosplice_vars}
\alias{are_exosplice_vars}
\title{Which variants are within exonic, splicing or exonic/splicing regions?}
\usage{
are_exosplice_vars(vars = NULL)
}
\arguments{
\item{vars}{The data frame of vars.}
}
\value{
a logical vector with \code{length} equal to \code{nrow(vars)}.
}
\description{
\code{are_exosplice_vars} reads in the region column from the vars file and
outputs a logical vector indicating which rows/variants have a region equal to
\code{"exonic"}, \code{"splicing"} or \code{"exonic;splicing"}.
}
\examples{
\dontrun{
big_logical_vec <- are_exosplice_vars(vars)
}
}
\seealso{
\code{\link[base]{Logic}}.
}

