% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/are_in_lr.R
\name{are_in_lr}
\alias{are_in_lr}
\title{Is the specific variant/row located within the provided linkage region?}
\usage{
are_in_lr(chr = NULL, startbp = NULL, endbp = NULL, vars = NULL)
}
\arguments{
\item{chr}{The linkage region chromosome. Takes values from 1 to 22
(autosomal) and X, Y, M (sex and mitochondrial).}

\item{startbp,endbp}{The start and end base pair coordinates.}

\item{vars}{The data frame of vars.}
}
\value{
A logical vector of length equal to nrow(vars), indicating if the row is
  within any of the linkage regions.
}
\description{
\code{are_in_lr} indicates which variants from the variant file lie within the
provided linkage region.
}
\examples{
\dontrun{
# assumes you have a vars data frame
big_logical_vec <- are_in_lr(chr = 4, startbp = 10000, endbp = 50000, vars)
}
}
\seealso{
\code{\link{Logic}}.

\code{\link{base::Logic}}.
}

