% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/are_rare_vars.R
\name{are_rare_vars}
\alias{are_rare_vars}
\title{Which variants are rare in the genomic databases?}
\usage{
are_rare_vars(vars = NULL, gendb = c("both_kg_esp", "kg", "esp"),
  maf = 0.05)
}
\arguments{
\item{vars}{The data frame of vars.}

\item{gendb}{The genomic database to choose (default: both 1KG and ESP6500).}
}
\value{
a logical vector with \code{length} equal to \code{nrow(vars)}.
}
\description{
\code{are_rare_vars} reads in the specified genomic database columns from the
vars file and outputs a logical vector indicating which rows/variants have a
MAF less than or equal to the specified maf, or are novel.
}
\examples{
\dontrun{
big_logical_vec <- are_rare_vars(vars, gendb = "both_kg_esp", maf = 0.05)
}
}
\seealso{
\code{\link[base]{Logic}}.
}

