% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_synsnv_vars.R
\name{are_synsnv_vars}
\alias{are_synsnv_vars}
\title{Which variants are synonymous SNVs?}
\usage{
are_synsnv_vars(vars = NULL)
}
\arguments{
\item{vars}{The data frame of vars.}
}
\value{
a logical vector with \code{length} equal to \code{nrow(vars)}.
}
\description{
\code{are_synsnv_vars} reads in the change column from the vars file and
outputs a logical vector indicating which rows/variants have a change equal to
\code{"synonymous SNV"}.
}
\examples{
\dontrun{
big_logical_vec <- are_synsnv_vars(vars)
}
}
\seealso{
\code{\link[base]{Logic}}.
}

