% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_genes.R
\name{clean_genes}
\alias{clean_genes}
\title{Cleans a vector of genes}
\usage{
clean_genes(gene_vec)
}
\arguments{
\item{gene_vec}{The atomic vector containing the annotated genes.}
}
\value{
character A clean version of the genes vector.
}
\description{
\code{clean_genes} reads in the gene variable from the variant file
and removes any additional info between parantheses.
}
\examples{
\dontrun{
clean_genes(vars$gene) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{gsub}}.
}

