% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_comma}
\alias{format_comma}
\title{Format numbers with commas for pretty printing}
\usage{
format_comma(x, trim = TRUE, ...)
}
\arguments{
\item{x}{The numeric vector.}

\item{trim}{logical; if FALSE, logical, numeric and complex values are
right-justified to a common width: if TRUE the leading blanks for
justification are suppressed (default: TRUE).}
}
\value{
A character version of x where thousands are separated by comma.
}
\description{
\code{format_comma} outputs the given numeric vector with a comma thousands
separator for pretty printing.
}
\examples{
\dontrun{
format_comma(vars$START)
format_comma(vars$END)
}
}
\seealso{
\code{\link{format}}
}

