% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_dec}
\alias{format_dec}
\title{Format floating point numbers with two decimals for pretty printing}
\usage{
format_dec(x, nsmall = 2, ...)
}
\arguments{
\item{x}{The numeric vector.}
}
\value{
A character version of x where only two decimals are used.
}
\description{
\code{format_dec} outputs the given numeric vector with two decimals
for pretty printing.
}
\examples{
\dontrun{
format_dec(vars$aaf.1KG)
format_dec(vars$esp6500_all)
}
}
\seealso{
\code{\link{format}}
}

