% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_tab.R
\name{gt_tab_per_fam}
\alias{gt_tab_per_fam}
\title{Genotype table per family.}
\usage{
gt_tab_per_fam(vars)
}
\arguments{
\item{vars}{The data frame of vars.}
}
\value{
A data frame with the families in rows and the different
genotypes in columns, and the counts of each genotype per family
in the cells.
}
\description{
\code{gt_tab_per_fam} reads in the genotype variables from the
variant file, groups them by individual family, and returns the number of
different genotypes in each of these groups.
}
\examples{
\dontrun{
gt_tab_per_fam(vars) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{table}}.
}

