% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_tab.R
\name{gt_tab_per_sample}
\alias{gt_tab_per_sample}
\title{Genotype table per sample.}
\usage{
gt_tab_per_sample(vars, rowSums = TRUE)
}
\arguments{
\item{vars}{The data frame of vars.}

\item{rowSums}{logical: should the sum of genotypes per
sample (row) be output (default: TRUE)?}
}
\value{
A data frame with the samples in rows and the different
genotypes in columns, and the counts of each genotype per sample
in the cells. A value of NA means that the sample did not have
the specific genotype.
}
\description{
\code{gt_tab_per_sample} reads in the genotype variables from the
variant file and returns the number of different genotypes in each of these
variables.
}
\examples{
\dontrun{
gt_tab_per_sample(vars, rowSums = TRUE) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{table}}.
}

