% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_tab.R
\name{is_gt_col}
\alias{is_gt_col}
\title{Which columns are GT columns?}
\usage{
is_gt_col(vars, pat = "_GT$")
}
\arguments{
\item{vars}{The data frame of vars.}

\item{pat}{character: the pattern which matches a typical genotype column
(default: "_GT$").}

\item{pat}{logical: should the sum of genotypes per sample (row) be output
(default: TRUE)?}
}
\value{
a logical vector where the ith TRUE element means that the ith column
is a genotype column.
}
\description{
\code{is_gt_col} uses \code{\link{grepl}} to check whether each column
name contains the string "_GT" or not.
}
\examples{
\dontrun{
is_gt_col(vars) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{grepl}}.
}

