% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gt_tab.R
\name{is_gt_col_for_fam}
\alias{is_gt_col_for_fam}
\title{Which columns are genotype columns for specified family?}
\usage{
is_gt_col_for_fam(vars, fam)
}
\arguments{
\item{vars}{The data frame of vars.}

\item{fam}{character: the family name}
}
\value{
a logical vector where the ith TRUE element means that the ith column
is a genotype column belonging to family <fam>.
}
\description{
\code{is_gt_col_for_fam} uses \code{\link{grepl}} to check whether each column
name contains the string "<fam>_GT" or not.
}
\examples{
\dontrun{
is_gt_col_for_fam(vars, "fam123") # assumes you have a vars data frame
}
}
\seealso{
\code{\link{grepl}}.
}

