% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_rare_vars.R
\name{is_rare_in}
\alias{is_rare_in}
\title{Check if this atomic vector is NA or less than a specified threshold.}
\usage{
is_rare_in(vec, thr)
}
\arguments{
\item{vec}{A numeric atomic vector.}

\item{thr}{The threshold.}
}
\value{
a logical vector with \code{length} equal to \code{length(vec)}.
}
\description{
\code{is_rare_in} outputs a logical vector indicating which elements have a
value less than or equal to the specified threshold, or are unavailable.
}
\seealso{
\code{\link{base::Logic}}.
\dontrun{
big_logical_vec <- is_rare_in(vec = vars$aaf.1KG, thr = 0.05)
}
}

