% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_tab.R
\name{maf_tab}
\alias{maf_tab}
\title{Returns a summary of a MAF variable (i.e. 1KG or ESP6500)}
\usage{
maf_tab(maf_vec)
}
\arguments{
\item{maf_vec}{The vector containing the MAF values.}
}
\value{
An integer vector with the counts of each condition
}
\description{
\code{maf_tab} reads in a MAF vector (i.e. 1KG or ESP6500) and outputs the
number of variants satisfying certain conditions.
}
\examples{
\dontrun{
maf_tab(vars$aaf.1KG) # assumes you have a vars data frame
maf_tab(vars$esp6500_all) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{sum}} and \code{\link{setNames}}.
}

