% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_collab_file}
\alias{read_collab_file}
\title{Reads in the raw merged collab file.}
\usage{
read_collab_file(fname, ...)
}
\arguments{
\item{fname}{The full file path to the merged_collab.txt file.}
}
\value{
The merged_collab.txt file as a data frame with \code{dplyr}'s tbl_df
  class.
}
\description{
\code{read_collab_file} reads in the merged_collab.txt file output by the MPS
pipeline. It is basically a wrapper around \code{read.delim}, but the
\code{na.strings} parameter is used to specify strings used throughout the
file to denote NAs.
}
\examples{
\dontrun{
full_path <- "~/Desktop/Fam1_merged_collab.txt"
DF <- read_collab_file(file.path(full_path))
}
}
\seealso{
\code{\link{read.table}}, \code{\link[dplyr]{tbl_df}} for more
  details.
}

