% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_collab_obj}
\alias{read_collab_obj}
\title{Reads in the raw merged collab file if stored as an R object (i.e. rds).}
\usage{
read_collab_obj(fname, ...)
}
\arguments{
\item{fname}{The full file path to the merged_collab.rds object.}
}
\value{
The merged_collab.txt file as a data frame with \code{dplyr}'s tbl_df
  class.
}
\description{
\code{read_collab_obj} reads in the merged_collab.txt file if it has already
been stored as a single R object. It is basically a wrapper around
\code{readRDS}, but it also wraps the object around \code{dplyr}'s tbl_df
class.
}
\examples{
\dontrun{
full_path <- "~/Desktop/Fam1_merged_collab.rds"
DF <- read_collab_obj(file.path(full_path))
}
}
\seealso{
\code{\link{readRDS}}, \code{\link[dplyr]{tbl_df}} for more
  details.
}

