% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_var_obj}
\alias{read_var_obj}
\title{Reads in a variant file stored as an rds object.}
\usage{
read_var_obj(fname, ...)
}
\arguments{
\item{fname}{The full file path to the variant object.}
}
\value{
The variant file as a data frame with \code{dplyr}'s tbl_df
  class.
}
\description{
\code{read_var_obj} reads in a variant file if it has already
been stored as a single R object. It is basically a wrapper around
\code{readRDS}, but it also wraps the object around \code{dplyr}'s tbl_df
class.
}
\examples{
\dontrun{
full_path <- "~/Desktop/Fam1_merged_collab.rds"
DF <- read_var_obj(file.path(full_path))
}
}
\seealso{
\code{\link{readRDS}}, \code{\link[dplyr]{tbl_df}} for more
  details.
}

