% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shorten.R
\name{shorten_change}
\alias{shorten_change}
\title{Shorten change column from variant file}
\usage{
shorten_change(change_vec)
}
\arguments{
\item{change_vec}{The character vector containing change values.}
}
\value{
A character vector with shortened change values.
}
\description{
\code{shorten_change} shortens the change column from the variant file in
order to print results in a more efficient manner.
}
\examples{
\dontrun{
tmp <- shorten_region(vars$change)
table(tmp, useNA = "ifany")
}
}
\seealso{
\code{\link[plyr]{mapvalues}}
}

