% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shorten.R
\name{shorten_pph2}
\alias{shorten_pph2}
\title{Shorten PPH2 column from variant file}
\usage{
shorten_pph2(pph2_vec)
}
\arguments{
\item{pph2_vec}{The character vector containing pph2 values.}
}
\value{
A character vector with shortened pph2 values.
}
\description{
\code{shorten_pph2} shortens the PPH2 column from the variant file in
order to print results in a more efficient manner.
}
\examples{
\dontrun{
tmp <- shorten_pph2(vars$pph2_pred)
table(tmp, useNA = "ifany")
}
}
\seealso{
\code{\link[plyr]{mapvalues}}
}

