% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten.R
\name{shorten_refalt}
\alias{shorten_refalt}
\title{Shorten REF/ALT column from variant file}
\usage{
shorten_refalt(x, max_len = 5)
}
\arguments{
\item{x}{The character vector containing REF/ALT values.}

\item{max_len}{The maximum number of characters used to print out the values.}
}
\value{
A character vector with shortened REF/ALT values.
A star '*' is appended as a flag.
}
\description{
\code{shorten_refalt} shortens the REF/ALT column from the variant file in
order to print results in a more efficient manner.
}
\examples{
\dontrun{
shorten_refalt(vars$REF)
shorten_refalt(vars$ALT)
}
}
\seealso{
\code{\link{strtrim}}
}

