% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shorten.R
\name{shorten_region}
\alias{shorten_region}
\title{Shorten region column from variant file}
\usage{
shorten_region(region_vec)
}
\arguments{
\item{region_vec}{The character vector containing region values.}
}
\value{
A character vector with shortened region values.
}
\description{
\code{shorten_region} shortens the region column from the variant file in
order to print results in a more efficient manner.
}
\examples{
\dontrun{
tmp <- shorten_region(vars$region)
table(tmp, useNA = "ifany")
}
}
\seealso{
\code{\link[plyr]{mapvalues}}
}

