% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shorten.R
\name{shorten_sift}
\alias{shorten_sift}
\title{Shorten SIFT column from variant file}
\usage{
shorten_sift(sift_vec)
}
\arguments{
\item{sift_vec}{The character vector containing sift values.}
}
\value{
A character vector with shortened sift values.
}
\description{
\code{shorten_sift} shortens the SIFT column from the variant file in
order to print results in a more efficient manner.
}
\examples{
\dontrun{
tmp <- shorten_sift(vars$SIFT.pred)
table(tmp, useNA = "ifany")
}
}
\seealso{
\code{\link[plyr]{mapvalues}}
}

