% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tidy_vars.R
\name{tidy_vars}
\alias{tidy_vars}
\title{Tidy up the vars data frame}
\usage{
tidy_vars(vars, samp_nm)
}
\arguments{
\item{vars}{The data frame of vars.}

\item{samp_nm}{The sample name to match (basically a string to match the
column names).}
}
\value{
The tidied vars data frame
}
\description{
\code{tidy_vars} optimises the vars data frame for printing i.e.
it shortens the various labels and separates thousands with comma etc.
}
\examples{
\dontrun{
tidy_vars(vars, "CANVAS886.*_GT$") # assumes you have a vars data frame
}
}
\seealso{
\code{\link[dplyr]{mutate}}
}

