% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{var_all_ok}
\alias{var_all_ok}
\title{Are the variable classes and names for the vars as expected?}
\usage{
var_all_ok(vars)
}
\arguments{
\item{vars}{The data frame of vars.}
}
\value{
A logical vector of length one.
}
\description{
\code{var_all_ok} is a convenience function that performs multiple varpr checks
related to the classes and names of variables in the variant file.
}
\examples{
\dontrun{
var_all_ok(vars) # assumes you have a vars data frame
}
}

