% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_columns.R
\name{var_cl_sample_ok}
\alias{var_cl_sample_ok}
\title{Are the sample var columns of the expected class?}
\usage{
var_cl_sample_ok(vars)
}
\arguments{
\item{vars}{The data frame of vars.}
}
\value{
A logical vector of length one.
}
\description{
\code{var_cl_sample_ok} checks that the sample columns of a variant file
are as expected.
}
\examples{
\dontrun{
var_cl_sample_ok(vars) # assumes you have a vars data frame
}
}

