% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_columns.R
\name{var_names_ok}
\alias{var_names_ok}
\title{Are the names of the main var columns as expected?}
\usage{
var_names_ok(vars)
}
\arguments{
\item{vars}{The data frame of vars.}
}
\value{
A logical vector of length one.
}
\description{
\code{var_names_ok} checks that the names of the main columns of a variant
file are as expected.
}
\examples{
\dontrun{
var_names_ok(vars) # assumes you have a vars data frame
}
}

