% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/counts.R
\name{vartype_count}
\alias{vartype_count}
\title{Counts the number of SNVs and INDELs (approximate estimation).}
\usage{
vartype_count(ref, alt)
}
\arguments{
\item{ref,alt}{The REF and ALT variables from the variant file.}
}
\value{
A data frame with the counts of SNVs and INDELs.
}
\description{
\code{vartype_count} reads in the REF and ALT variables from the variant file
and counts how many variants "rows" there are with a single base. This should
imply a SNV, although exceptions to this are quite possible (can't think of an
example right now).
}
\examples{
\dontrun{
vartype_count(vars$REF, vars$ALT) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{grep}}
}

