% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vartype_count.R
\name{vartype_count_per_chr}
\alias{vartype_count_per_chr}
\title{Counts the number of SNVs and INDELs within each chromosome}
\usage{
vartype_count_per_chr(vars)
}
\arguments{
\item{vars}{The vars file that contains the REF and ALT variables.}
}
\value{
A data frame with the counts of SNVs and INDELs.
}
\description{
\code{vartype_count_per_chr} reads in the REF and ALT variables from the variant file
and counts how many variant "rows" there are with a single base. This should
imply a SNV, although exceptions to this are quite possible (can't think of an
example right now).
}
\examples{
\dontrun{
vartype_count_per_chr(vars) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{vartype_count}}.
}

