% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/counts.R
\name{gt_tab_per_sample}
\alias{gt_tab_per_sample}
\title{Returns the number of different genotypes per sample.}
\usage{
gt_tab_per_sample(vars, rowSums = TRUE)
}
\arguments{
\item{vars}{The data frame of vars.}

\item{rowSums}{logical: should the sum of genotypes per sample (row) be output (default: TRUE)?}
}
\value{
A data frame with the counts of each genotype. NA means the sample did not have the specific genotype.
}
\description{
\code{gt_tab_per_sample} reads in the genotype variables from the
variant file and returns the number of different genotypes in each of these
variables.
}
\examples{
\dontrun{
gt_tab_per_sample(vars, rowSums = TRUE) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{table}}.
}

