% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_gt_col.R
\name{is_gt_col}
\alias{is_gt_col}
\title{Which columns are genotype columns?}
\usage{
is_gt_col(vars, pat = "_GT$")
}
\arguments{
\item{vars}{The data frame of vars.}

\item{pat}{character: the pattern which matches a typical genotype column
(default: "_GT$").}
}
\value{
an atomic vector of the indices of the columns that yielded a match.
}
\description{
\code{is_gt_col} uses \code{\link{grep}} to check which columns of the vars
data frame contain the string "_GT".
}
\examples{
\dontrun{
is_gt_col(vars) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{grep}}.
}

