% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_gt_col.R
\name{is_gt_col_for_fam}
\alias{is_gt_col_for_fam}
\title{Which columns are genotype columns for the specified family?}
\usage{
is_gt_col_for_fam(vars, fam)
}
\arguments{
\item{vars}{The data frame of vars.}

\item{fam}{character: the family name}
}
\value{
an atomic vector of the indices of the vars columns that yielded a match.
}
\description{
\code{is_gt_col_for_fam} uses \code{\link{grep}} to check which columns of
the vars data frame contain the string "<fam>_GT".
}
\examples{
\dontrun{
is_gt_col_for_fam(vars, "fam123") # assumes you have a vars data frame
}
}
\seealso{
\code{\link{grep}}.
}

