% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype_count.R
\name{vartype_count}
\alias{vartype_count}
\title{Counts the number of SNVs and INDELs.}
\usage{
vartype_count(vars)
}
\arguments{
\item{vars}{The vars file that contains the REF and ALT variables.}
}
\value{
A data frame with the counts of SNVs, INDELs and Total.
}
\description{
\code{vartype_count} reads in the REF and ALT variables from the variant file
and counts how many variant "rows" there are with a single base. This should
imply a SNV, although exceptions to this are quite possible (can't think of an
example right now).
}
\examples{
\dontrun{
vartype_count(vars) # assumes you have a vars data frame
}
}
\seealso{
\code{\link{vartype_count_per_chr}}.
}

